requirements.txt
python-telegram-bot==21.7
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2
pillow==10.1.0
opencv-python==4.9.0.80
pytesseract==0.3.10
easyocr==1.7.1
scikit-learn==1.4.2
python-dotenv==1.0.0
